Experiment with explicitly throwing
I was going to attempt to essentially use spec to enforce preconditions and
postconditions and thereby mostly avoid "normal" exception-handling patterns.
This would give me some advantages, including the ability to get some better
reporting. But, mostly because I wanted to see how well it would work.

Unfortunately the reporting is terrible! The reason why it's terrible is that
it's unreadable, because I pass around the system map, which is essentially the
whole state, including lots of stuff an individual function might not
explicitly need at the moment. So you can't read anything because it dumps out
the input parameters, and one of the input parameters is the entire state of
the game.

That isn't really an indictment of spec, it just flat-out doesn't work well
with this design but hey, this is a pretty specific design (stateless ECS), to
be fair. But it does mean it doesn't seem to work well.

Heh, I'm mostly writing this because I was sick for, like, a week, and I came
back to look at this and I saw the error messages I was working with and went
"Wow how do you read this, it's just pages of noise with 2 interesting data
points. Should I...search the error text for - wait this is clearly not, like,
working."